Add endpoint to Federator API, cleanup
